converter: support ServiceMonitor's serviceDiscoveryRole

+++ api/operator/v1beta1/vmstaticscrape_types.go
@@ -19,11 +19,11 @@ type VMStaticScrapeSpec struct {
 	TargetEndpoints []*TargetEndpoint `json:"targetEndpoints"`
 	// SampleLimit defines per-scrape limit on number of scraped samples that will be accepted.
 	// +optional
-	SampleLimit uint64 `json:"sampleLimit,omitempty"`
+	SampleLimit int `json:"sampleLimit,omitempty"`
 	// SeriesLimit defines per-scrape limit on number of unique time series
 	// a single target can expose during all the scrapes on the time window of 24h.
 	// +optional
-	SeriesLimit uint64 `json:"seriesLimit,omitempty"`
+	SeriesLimit int `json:"seriesLimit,omitempty"`
 	// ScrapeClass defined scrape class to apply
 	// +optional
 	ScrapeClassName *string `json:"scrapeClass,omitempty"`
